feat: Make error boundary pages consistent to the theme
