--- apps/web/app/(error)/error/layout.tsx
@@ -1,43 +1,0 @@
-import type { Metadata } from "next"
-
-import "@/app/globals.css"
-import QueryClientWrapper from "@/common/components/QueryClientWrapper"
-import GlobalModalWrapper from "@/common/components/GlobalModalWrapper"
-import { Toaster } from "react-hot-toast"
-import React from "react"
-import { LOGO_SINGLE_IMAGE } from "@/common/constants/index"
-import { APP_NAME } from "@repo/constants"
-import Header from "@/common/components/Header"
-import Footer from "@/common/components/Footer"
-import AuthStateProvider from "@/common/components/AuthStateProvider"
-
-export const metadata: Metadata = {
-  title: APP_NAME,
-  description: `Generated by ${APP_NAME}`,
-}
-
-export default async function RootLayout({
-  children,
-}: {
-  readonly children: React.ReactNode
-}) {
-  return (
-    <html lang="en">
-      <link rel="icon" type="image/x-icon" href={LOGO_SINGLE_IMAGE} />
-      <body>
-        <Toaster />
-        <QueryClientWrapper>
-          <AuthStateProvider>
-            <GlobalModalWrapper>
-              <div className="flex flex-col gap-2">
-                <Header />
-              </div>
-              <div>{children}</div>
-              <Footer />
-            </GlobalModalWrapper>
-          </AuthStateProvider>
-        </QueryClientWrapper>
-      </body>
-    </html>
-  )
-}
