--- apps/web/app/(error)/error/page.tsx
@@ -1,7 +1,0 @@
-import GlobalError from "@/module/Error/global-error"
-
-const NotFoundPage = () => {
-  return <GlobalError />
-}
-
-export default NotFoundPage
