+++ apps/web/app/layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from "next"
+
+import "@/app/globals.css"
+import React from "react"
+import { LOGO_SINGLE_IMAGE } from "@/common/constants/index"
+import { APP_NAME } from "@repo/constants"
+
+export const metadata: Metadata = {
+  title: APP_NAME,
+  description: `Generated by ${APP_NAME}`,
+}
+
+export default async function RootLayout({
+  children,
+}: {
+  readonly children: React.ReactNode
+}) {
+  return (
+    <html lang="en">
+      <link rel="icon" type="image/x-icon" href={LOGO_SINGLE_IMAGE} />
+      <body>
+        {children}
+      </body>
+    </html>
+  )
+}
